requirements.txt

django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

channels==4.0.0
daphne==4.1.2
channels-redis==4.2.0
websockets==11.0.3

celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

psycopg2-binary==2.9.9
django-environ==0.11.2
python-dotenv==1.0.1

pytest
pytest-django==4.8.0

flower
daphne


pytest-asyncio==0.26.0

